Refactor kcidb.orm module for improved readability and maintainability
Extract parts of kcidb.orm module into separate modules for improved
maintainability and readability. Specifically, we refactored the module by
creating two new modules - kcidb.orm.data for SCHEMA and dependencies, and
kcidb.orm.query for Pattern and dependencies. Updated all relevant
references to the new modules. No changes in behavior.
