+++ kcidb/db/sqlite/v04_00.py
@@ -646,13 +646,13 @@ class Schema(AbstractSchema):
         Render a pattern for raw OO data into a query.
 
         Args:
-            pattern:    The pattern (instance of kcidb.orm.Pattern) to
+            pattern:    The pattern (instance of kcidb.orm.query.Pattern) to
                         render.
 
         Returns:
             The SQL query string and the query parameters.
         """
-        assert isinstance(pattern, orm.Pattern)
+        assert isinstance(pattern, orm.query.Pattern)
         obj_type = pattern.obj_type
         type_query_string = cls.OO_QUERIES[obj_type.name]["statement"]
         if pattern.obj_id_set:
@@ -713,18 +713,18 @@ class Schema(AbstractSchema):
         Query raw object-oriented data from the database.
 
         Args:
-            pattern_set:    A set of patterns ("kcidb.orm.Pattern"
+            pattern_set:    A set of patterns ("kcidb.orm.query.Pattern"
                             instances) matching objects to fetch.
         Returns:
             A dictionary of object type names and lists containing retrieved
             objects of the corresponding type.
         """
         assert isinstance(pattern_set, set)
-        assert all(isinstance(r, orm.Pattern) for r in pattern_set)
+        assert all(isinstance(r, orm.query.Pattern) for r in pattern_set)
 
         # Render all queries for each type
         obj_type_queries = {}
-        for obj_type in orm.SCHEMA.types.values():
+        for obj_type in orm.data.SCHEMA.types.values():
             for pattern in pattern_set:
                 # TODO: Avoid adding the same patterns multiple times
                 if pattern.obj_type == obj_type:
@@ -765,7 +765,7 @@ class Schema(AbstractSchema):
             finally:
                 cursor.close()
 
-        assert LIGHT_ASSERTS or orm.SCHEMA.is_valid(objs)
+        assert LIGHT_ASSERTS or orm.data.SCHEMA.is_valid(objs)
         return objs
 
     def load(self, data):
